+++ Mac-Health-Check.zsh
@@ -20,6 +20,7 @@
 # Version 3.0.0, 24-Sep-2025, Dan K. Snelson (@dan-snelson)
 #   - First (attempt at a) MDM-agnostic release
 #   - Added "System Memory" and "System Storage" capacity information (Pull Request #36; thanks again, @HowardGMac!)
+#   - Corrected misspelling of "Certificate" in multiple locations (Pull Request #41; thanks, @HowardGMac!)
 #   - Introduces an `operationMode` of "Silent" to run all checks and log results without displaying a dialog to the user
 #
 ####################################################################################################
@@ -35,7 +36,7 @@
 export PATH=/usr/bin:/bin:/usr/sbin:/sbin:/usr/local/bin/
 
 # Script Version
-scriptVersion="3.0.0b26"
+scriptVersion="3.0.0b27"
 
 # Client-side Log
 scriptLog="/var/log/org.churchofjesuschrist.log"