+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### 3.0.0 (24-Sep-2025) 
 - First (attempt at a) MDM-agnostic release
 - Added "System Memory" and "System Storage" capacity information (Pull Request #36; thanks again, @HowardGMac!)
+- Corrected misspelling of "Certificate" in multiple locations (Pull Request #41; thanks, @HowardGMac!)
 - Introduces a new `operationMode` of "Silent" to run all checks and log results without displaying a dialog to the end-user
 
     > :warning: **Breaking Change** :warning: